Be verbose for single slave.
